Add files via upload
Snake Game is a simple implementation of the traditional snake game. The program is developed using the Java language and the JavaFX library to create a graphical user interface. It allows the user to play with a small snake and increase its length by eating food. Points are recorded and displayed on the screen. The program contains two game modes, and the user controls the movement of the snake through the steering buttons. It checks if the snake hits itself or edges to end the game and displays the final score and player score. The program includes sound effects for various events in the game. The program provides an enjoyable experience of the traditional snake game with a simple graphical interface and sound effects.